fix(ci): python-full-coverage — bitbucket_client.py crossed the 1500-line guardrail
The base-ref-drift fix (1344e024e) added four lines to BitbucketClient
(destination_branch + PRStatus.base_ref), taking src/awf/common/bitbucket_client.py
from exactly 1500 to 1504 lines and failing
test_first_party_code_files_stay_under_line_limit in the coverage shard.

Move the D2 HTTP transport core — _request/_request_json/_request_text, cursor
pagination, the forge-origin SSRF guard, the 429/near-limit backoff, ETag LRU, and
redacted error mapping — verbatim onto a _BitbucketHttpMixin in a new
bitbucket_client_http module, following the existing
bitbucket_client_errors/_parsing/_paths split. Behavior is unchanged: the mixin
declares the instance attributes BitbucketClient.__init__ already sets, and the
class keeps the same public surface, so the moved lines stay covered by the same
tests and the aggregate coverage total is unchanged. bitbucket_client.py is now
1133 lines, with real headroom instead of a one-line margin.

Focused checks: ruff check/format and mypy on both modules;
tests/unit/test_core_decomposition_maintainability.py,
tests/unit/common/test_bitbucket_client_parts,
tests/unit/common/test_bitbucket_client_forge.py, tests/unit/common/test_forge*.py,
tests/unit/runtime/test_pr_monitor_runner_parts, test_pr_creator.py,
test_pr_monitor_merge_failures.py, test_pr_monitor_merge_methods.py — all green.
Full validation and the coverage gate are AWF/CI-owned after agent completion.
